requirements.txt
# Weed Detection CNN-ViT Hybrid
# ================================
# Training Environment (Kaggle / Cloud GPU)
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
torchmetrics>=1.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=9.5.0
grad-cam>=1.4.0
onnx>=1.14.0
onnxruntime-gpu>=1.15.0
PyYAML>=6.0

# Jetson Nano Inference Only
# onnxruntime-gpu>=1.15.0
# numpy>=1.24.0
# Pillow>=9.5.0
# opencv-python>=4.7.0
